+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog
  * optimized `Valid::andX()` and `Valid::orX()`
  * optimized `Invalid::andX()` and `Invalid::orX()`
  * removed `true()` and `false()`. Use `same()` instead
+ * renamed `valid()` to `true()` and `invalid()` to `false()`
 
 * 1.0.0-beta2 (2015-04-13)